Correct recurring, httpclientTimeout to httpRequestTimeout, empty string check
