+++ Adyen/Config.cs
@@ -12,7 +12,7 @@ namespace Adyen
         public Environment Environment { get; set; }
         public string Endpoint { get; set; }
         public string ApplicationName { get; set; }
-        public int HttpClientTimeout { get; set; }
+        public int HttpRequestTimeout { get; set; }
         //HPP specific
         public string HppEndpoint { get; set; }
         public string SkinCode{ get; set; }
