+++ Adyen/Model/Recurring/RecurringDetailsResult.cs
@@ -15,7 +15,7 @@ using System.ComponentModel.DataAnnotations;
 using System.Runtime.Serialization;
 using System.Text;
 
-namespace Adyen.Model.Reccuring
+namespace Adyen.Model.Recurring
 {
     /// <summary>
     /// RecurringDetailsResult
